Restructure frontend directory
This commit is the first part of refactoring the frontend and backend
directories into a single npm package.

Specifically, the frontend directory's package.json and other config
files are moved to the root directory. Components and frontend services
remain in the frontend directory, which is placed inside a `src` folder
where the backend directory will also be moved later.

Finally, I changed the package name to "blogapp" in `package.json` and
`package-lock.json` files.